+++ src/theme/components/featuredContent.ts
@@ -2,6 +2,7 @@ import { createMultiStyleConfigHelpers } from "@chakra-ui/styled-system";
 import { StyleFunctionProps } from "@chakra-ui/system";
 import { wrapperStyles } from "./global";
 import { screenreaderOnly } from "./globalMixins";
+import { setContainerStyles } from "../../utils/setContainerStyles";
 
 // This function creates a set of function that helps us
 // create multipart component styles.
@@ -46,18 +47,48 @@ const FeaturedContent = defineMultiStyleConfig({
             bgColor: "dark.ui.bg.default",
           },
           img: screenreaderOnly(),
+          containerType: "inline-size",
+          ...setContainerStyles({
+            breakpoint: "base",
+            styles: {
+              "[data-wrapper]": {
+                flexDirection: imageAtEnd ? "column-reverse" : "column",
+              },
+              "[data-imageWrapper]": {
+                height: "320px",
+                width: "100%",
+              },
+            },
+          }),
+          ...setContainerStyles({
+            breakpoint: "md",
+            styles: {
+              "[data-wrapper]": {
+                flexDirection: imageAtEnd ? "row-reverse" : "row",
+                paddingLeft: full ? "s" : null,
+                paddingRight: full ? "s" : null,
+              },
+              "[data-imageWrapper]": {
+                height: "auto",
+                width: wrapperWidth,
+              },
+              "[data-text]": {
+                /** The `paddingLeft` attribute is used to adjust the spacing around the
+                 * text when the image is positioned at the end. For aesthetic reasons,
+                 * we opted to not adjust the spacing around the text when the image is
+                 * positioned at the start.
+                 * */
+                paddingStart: full && imageAtEnd ? 0 : null,
+              },
+            },
+          }),
         },
         wrapper: {
           ...wrapperStyles,
           alignItems: "stretch",
           display: "flex",
-          flexDirection: imageAtEnd
-            ? { base: "column-reverse", md: "row-reverse" }
-            : { base: "column", md: "row" },
           maxWidth: full ? "1280px" : "100%",
           minHeight: "320px",
-          paddingLeft: full ? { base: null, md: "s" } : null,
-          paddingRight: full ? { base: null, md: "s" } : null,
         },
         text: {
           display: "flex",
@@ -65,18 +96,10 @@ const FeaturedContent = defineMultiStyleConfig({
           flexDirection: "column",
           justifyContent: "center",
           padding: "l",
-          /** The `paddingLeft` attribute is used to adjust the spacing around the
-           * text when the image is positioned at the end. For aesthetic reasons,
-           * we opted to not adjust the spacing around the text when the image is
-           * positioned at the start.
-           * */
-          paddingStart: full && imageAtEnd ? { base: null, md: 0 } : null,
         },
         imgWrapper: {
           backgroundPosition: "center",
           backgroundSize: "cover",
-          height: { base: "320px", md: "auto" },
-          width: { base: "100%", md: wrapperWidth },
         },
       };
     }