+++ src/components/Modal/modalChangelogData.ts
@@ -10,6 +10,13 @@ import { ChangelogData } from "../../utils/ComponentChangelogTable";
 
 export const changelogData: ChangelogData[] = [
   {
+    date: "Prerelease",
+    version: "Prerelease",
+    type: "Update",
+    affects: ["Styles"],
+    notes: ["Updates spacing and border radius to match VDL."],
+  },
+  {
     date: "2024-05-15",
     version: "3.1.3",
     type: "Update",