+++ src/components/Fieldset/fieldsetChangelogData.ts
@@ -10,6 +10,15 @@ import { ChangelogData } from "../../utils/ComponentChangelogTable";
 
 export const changelogData: ChangelogData[] = [
   {
+    date: "Prerelease",
+    version: "Prerelease",
+    type: "Update",
+    affects: ["Styles"],
+    notes: [
+      'Changes the font weight for the `(required)` label to `"medium"`.',
+    ],
+  },
+  {
     date: "2024-09-19",
     version: "3.3.2",
     type: "Update",