+++ CHANGELOG.md
@@ -8,6 +8,15 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ## Prerelease
 
+### Adds
+
+- Adds the `Container Query Guide` page to the `Development Guide` section of Storybook.
+
+### Updates
+
+- Updates the spacing between heading and content in `Banner` component.
+- Updates the `Card`, `FeaturedContent`, and `SearchBar` components to use container queries.
+
 ## 3.5.4 (February 13, 2025)
 
 ### Adds