+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 - Updates the spacing on `Modal`.
 - Updates the spacing between heading and content in `Banner` component.
 - Updates the `Card`, `FeaturedContent`, and `SearchBar` components to use container queries.
+- Updates `Image` to include 'fourByOne' and 'twoByThree' aspect ratio options.
 
 ## 3.5.4 (February 13, 2025)