+++ CHANGELOG.md
@@ -8,9 +8,18 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ## Prerelease
 
+### Adds
+
+- Adds the `Container Query Guide` page to the `Development Guide` section of Storybook.
+
 ### Updates
 
+- Updates the spacing on `Modal`.
 - Updates the spacing between heading and content in `Banner` component.
+- Updates the `Card`, `FeaturedContent`, and `SearchBar` components to use container queries.
+- Updated the `"filter"` variant of `TagSet` to remove tag button wrapper when `isDismissible` and `onClick` are false.
+- Updates `Image` to include 'fourByOne' and 'twoByThree' aspect ratio options.
+
 
 ### Fixes
 
@@ -21,6 +30,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 ### Adds
 
 - Adds the `"utilityCheck"` option to the `Icon` component.
+- Adds the `getTextFromElement` utility function.
 
 ### Updates
 
@@ -32,6 +42,9 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 - Updates the `FeedbackBox` component to set the font weight for the popup header to `medium`.
 - Updates the `primary` variant of the `Hero` component to have appropriate padding for desktop and mobile, and `max-width` of 860px
 - Updates the `campaign` variant of the `Hero` component to have consistent padding.
+- Updates the `Label` component to set the font weight for the `(required)` label to `"medium"`.
+- Updates the `Fieldset` component to set the font weight for the `(required)` label to `"medium"`.
+- Updates the `FeedbackBox` component to set the font weight for the `(required)` label on the `Textarea` element to `"medium"`.
 
 ### Fixes